requirements.txt
pillow==9.4.0
pillow-heif==0.9.1
exifread==3.0.0
ffmpeg-python==0.2.0